Fix lossy conversion of Content-Length
Actual variable holding content length is int64_t, but it was assigned
by explicit cast to size_t. On 32-bit systems it's a lossy conversion,
so it was removed altogether.

(cherry picked from commit 50b8277c4ed38defdc6d68a93ab410fe16a23e4e)

+++ plugins/curl/curl.c
@@ -412,7 +412,7 @@ curl_open (int readonly)
     goto err;
   }
 
-  h->exportsize = (size_t) d;
+  h->exportsize = d;
 #endif
   nbdkit_debug ("content length: %" PRIi64, h->exportsize);
